Re-added the bit more detailed logging output of the input sentence.
Our bimachines would now be perfect, if we used regular a
(character-based) alphabet. But with these symbols, we are a bit
unlucky: the left-right sequential transduction loses the information
encoded in the identity Symbols, i.e. the actual text token associated
with the symbol. Let's think a bit how we could preserve this information...

git-svn-id: https://svn.code.sf.net/p/apertium/svn/branches/fomacg_bimachines_pristine@57215 72bbbca6-d526-0410-a7d9-f06f51895060